Rename rich-text component in kebab case
