+++ src/ggrc-client/js/components/form/fields/rich-text-form-field.js
@@ -3,7 +3,7 @@
  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
  */
 
-import '../../rich_text/rich_text';
+import '../../rich-text/rich-text';
 import template from './rich-text-form-field.stache';
 
 export default can.Component.extend({
