+++ src/ggrc-client/js/components/rich-text/rich-text.stache
