+++ src/ggrc-client/js/controllers/modals/modals_controller.js
@@ -27,7 +27,7 @@ import '../../components/modal_wrappers/assessment_template_form';
 import '../../components/autocomplete/autocomplete';
 import '../../components/external-data-autocomplete/external-data-autocomplete';
 import '../../components/person/person-data';
-import '../../components/rich_text/rich_text';
+import '../../components/rich-text/rich-text';
 import '../../components/modal_wrappers/checkboxes_to_list';
 import '../../components/deferred-mapper';
 import '../../components/modal_wrappers/assessment-modal';
