+++ src/ggrc-client/js/entrypoints/dashboard/index.js
@@ -25,7 +25,7 @@ import '../../components/tabs/tab-container';
 import '../../components/revision-log/revision-log';
 import '../../components/quick_form/quick_update';
 import '../../components/reminder';
-import '../../components/rich_text/rich_text';
+import '../../components/rich-text/rich-text';
 import '../../components/object_cloner/object_cloner';
 import '../../components/tree_pagination/tree_pagination';
 import '../../components/tree/tree-widget-container';
